+++ packages/strapi-admin/config/functions/bootstrap.js
@@ -20,7 +20,7 @@ module.exports = async () => {
   await strapi.admin.services.permission.ensureBoundPermissionsInDatabase();
   await strapi.admin.services.user.migrateUsers();
   await strapi.admin.services.role.createRolesIfNoneExist();
-  await strapi.admin.services.permission.resetSuperAdminPermissions();
+  await strapi.admin.services.role.resetSuperAdminPermissions();
   await strapi.admin.services.role.displayWarningIfNoSuperAdmin();
   await strapi.admin.services.user.displayWarningIfUsersDontHaveRole();
 };
