+++ packages/strapi-admin/services/__tests__/permission.test.js
@@ -19,86 +19,6 @@ describe('Permission Service', () => {
     });
   });
 
-  describe('Assign permissions', () => {
-    test('Delete previous permissions', async () => {
-      const createMany = jest.fn(() => Promise.resolve([]));
-      const getSuperAdmin = jest.fn(() => Promise.resolve({ id: 0 }));
-      const sendDidUpdateRolePermissions = jest.fn();
-      const find = jest.fn(() => Promise.resolve([{ id: 3 }]));
-      const deleteFn = jest.fn();
-      const getAll = jest.fn(() => []);
-
-      global.strapi = {
-        admin: {
-          services: {
-            metrics: { sendDidUpdateRolePermissions },
-            permission: { actionProvider: { getAll } },
-            role: { getSuperAdmin },
-          }
-        },
-        query() {
-          return { delete: deleteFn, createMany, find };
-        },
-      };
-
-      await permissionService.assign(1, []);
-
-      expect(deleteFn).toHaveBeenCalledWith({ id_in: [3] });
-    });
-
-    test('Create new permissions', async () => {
-      const permissions = Array(5)
-        .fill(0)
-        .map((v, i) => ({ action: `action-${i}` }));
-
-      const deleteFn = jest.fn(() => Promise.resolve([]));
-      const createMany = jest.fn(() => Promise.resolve([]));
-      const getSuperAdmin = jest.fn(() => Promise.resolve({ id: 0 }));
-      const sendDidUpdateRolePermissions = jest.fn();
-      const find = jest.fn(() => Promise.resolve([]));
-      const getAll = jest.fn(() => permissions.map(perm => ({ actionId: perm.action })));
-      const removeUnkownConditionIds = jest.fn(conds => _.intersection(conds, ['cond']));
-
-      global.strapi = {
-        admin: {
-          services: {
-            metrics: { sendDidUpdateRolePermissions },
-            role: { getSuperAdmin },
-            permission: {
-              actionProvider: { getAll },
-              conditionProvider: {
-                getAll: jest.fn(() => [{ id: 'admin::is-creator' }]),
-              },
-            },
-            condition: {
-              removeUnkownConditionIds,
-            },
-          },
-        },
-        query() {
-          return { delete: deleteFn, createMany, find };
-        },
-      };
-
-      const permissionsToAssign = [...permissions];
-      permissionsToAssign[4] = {
-        ...permissions[4],
-        conditions: ['cond', 'unknown-cond'],
-      };
-
-      await permissionService.assign(1, permissionsToAssign);
-
-      expect(createMany).toHaveBeenCalledTimes(1);
-      expect(createMany).toHaveBeenCalledWith([
-        { action: 'action-0', conditions: [], fields: null, role: 1, subject: null },
-        { action: 'action-1', conditions: [], fields: null, role: 1, subject: null },
-        { action: 'action-2', conditions: [], fields: null, role: 1, subject: null },
-        { action: 'action-3', conditions: [], fields: null, role: 1, subject: null },
-        { action: 'action-4', conditions: ['cond'], fields: null, role: 1, subject: null },
-      ]);
-    });
-  });
-
   describe('Find User Permissions', () => {
     test('Find calls the right db query', async () => {
       const find = jest.fn(({ role_in }) => role_in);
@@ -227,72 +147,6 @@ describe('Permission Service', () => {
 
       // restauring actionProvider
       permissionService.actionProvider.getAllByMap = prevGetAllByMap;
-    });
-  });
-
-  describe('resetSuperAdminPermissions', () => {
-    test('No superAdmin role exist', async () => {
-      const getSuperAdmin = jest.fn(() => Promise.resolve(undefined));
-      const createMany = jest.fn();
-
-      global.strapi = {
-        query: () => ({ createMany }),
-        admin: { services: { role: { getSuperAdmin } } },
-      };
-
-      await permissionService.resetSuperAdminPermissions();
-
-      expect(createMany).toHaveBeenCalledTimes(0);
-    });
-    test('Reset super admin permissions', async () => {
-      const actions = [
-        {
-          actionId: 'action-1',
-          subjects: ['country'],
-          section: 'contentTypes',
-        },
-      ];
-      const permissions = [
-        {
-          action: 'action-1',
-          subject: 'country',
-          fields: ['name'],
-          conditions: [],
-        },
-      ];
-      const getAll = jest.fn(() => actions);
-      const getAllConditions = jest.fn(() => []);
-      const find = jest.fn(() => [{ action: 'action-2', id: 2 }]);
-      const deleteFn = jest.fn(() => []);
-      const getPermissionsWithNestedFields = jest.fn(() => [...permissions]); // cloned, otherwise it is modified inside resetSuperAdminPermissions()
-      const getSuperAdmin = jest.fn(() => Promise.resolve({ id: 1 }));
-      const createMany = jest.fn(() => Promise.resolve([{ ...permissions[0], role: { id: 1 } }]));
-      const removeUnkownConditionIds = jest.fn(conds => conds);
-
-      global.strapi = {
-        query: () => ({ createMany, find, delete: deleteFn }),
-        admin: {
-          services: {
-            permission: {
-              actionProvider: { getAll },
-              conditionProvider: { getAll: getAllConditions },
-            },
-            condition: { removeUnkownConditionIds },
-            'content-type': { getPermissionsWithNestedFields },
-            role: { getSuperAdmin },
-          },
-        },
-      };
-
-      await permissionService.resetSuperAdminPermissions();
-
-      expect(deleteFn).toHaveBeenCalledWith({ id_in: [2] });
-      expect(createMany).toHaveBeenCalledWith([
-        {
-          ...permissions[0],
-          role: 1,
-        },
-      ]);
     });
   });
 });
