+++ packages/strapi-admin/services/permission.js
@@ -3,7 +3,6 @@
 const _ = require('lodash');
 const { flatMap, filter } = require('lodash/fp');
 const pmap = require('p-map');
-const { validatePermissionsExist } = require('../validation/permission');
 const createPermissionsManager = require('./permission/permissions-manager');
 const createConditionProvider = require('./permission/condition-provider');
 const createPermissionEngine = require('./permission/engine');
@@ -14,20 +13,6 @@ const { createPermission } = require('../domain/permission');
 
 const conditionProvider = createConditionProvider();
 const engine = createPermissionEngine(conditionProvider);
-
-const fieldsToCompare = ['action', 'subject', 'fields', 'conditions'];
-const getPermissionWithSortedFields = perm => {
-  const sortedPerm = _.cloneDeep(perm);
-  if (Array.isArray(sortedPerm.fields)) {
-    sortedPerm.fields.sort();
-  }
-  return sortedPerm;
-};
-const arePermissionsEqual = (perm1, perm2) =>
-  _.isEqual(
-    _.pick(getPermissionWithSortedFields(perm1), fieldsToCompare),
-    _.pick(getPermissionWithSortedFields(perm2), fieldsToCompare)
-  );
 
 /**
  * Removes unwanted fields from a permission
@@ -62,9 +47,7 @@ const deleteByIds = ids => {
  * @param permissions
  * @returns {Promise<*[]|*>}
  */
-const createMany = async (roleId, permissions) => {
-  permissions.forEach(p => p.role = roleId);
-
+const createMany = async permissions => {
   return strapi.query('permission', 'admin').createMany(permissions);
 };
 
@@ -85,57 +68,6 @@ const update = async (params, attributes) => {
  */
 const find = (params = {}) => {
   return strapi.query('permission', 'admin').find(params, []);
-};
-
-/**
- * Assign permissions to a role
- * @param {string|int} roleId - role ID
- * @param {Array<Permission{action,subject,fields,conditions}>} permissions - permissions to assign to the role
- */
-const assign = async (roleId, permissions = []) => {
-  try {
-    await validatePermissionsExist(permissions);
-  } catch (err) {
-    throw strapi.errors.badRequest('ValidationError', err);
-  }
-
-  const superAdmin = await strapi.admin.services.role.getSuperAdmin();
-  const isSuperAdmin = superAdmin && superAdmin.id === roleId;
-
-  const permissionsWithRole = permissions.map(permission =>
-    createPermission({
-      ...permission,
-      conditions: strapi.admin.services.condition.removeUnkownConditionIds(permission.conditions),
-      role: roleId,
-    })
-  );
-
-  const existingPermissions = await find({ role: roleId, _limit: -1 });
-  const permissionsToAdd = _.differenceWith(
-    permissionsWithRole,
-    existingPermissions,
-    arePermissionsEqual
-  );
-  const permissionsToDelete = _.differenceWith(
-    existingPermissions,
-    permissionsWithRole,
-    arePermissionsEqual
-  );
-  const permissionsToReturn = _.differenceBy(existingPermissions, permissionsToDelete, 'id');
-
-  if (permissionsToDelete.length > 0) {
-    await deleteByIds(permissionsToDelete.map(p => p.id));
-  }
-  if (permissionsToAdd.length > 0) {
-    const createdPermissions = await createMany(roleId, permissionsToAdd);
-    permissionsToReturn.push(...createdPermissions.map(p => ({ ...p, role: p.role.id })));
-  }
-
-  if (!isSuperAdmin && (permissionsToAdd.length || permissionsToDelete.length)) {
-    await strapi.admin.services.metrics.sendDidUpdateRolePermissions();
-  }
-
-  return permissionsToReturn;
 };
 
 /**
@@ -250,41 +182,9 @@ const ensureBoundPermissionsInDatabase = async () => {
           fields: BOUND_ACTIONS_FOR_FIELDS.includes(action) ? fields : null,
         })
       );
-      await createMany(editorRole.id, permissions);
+      await createMany(permissions);
     }
   }
-};
-
-/**
- * Reset super admin permissions (giving it all permissions)
- * @returns {Promise<>}
- */
-const resetSuperAdminPermissions = async () => {
-  const superAdminRole = await strapi.admin.services.role.getSuperAdmin();
-  if (!superAdminRole) {
-    return;
-  }
-
-  const allActions = strapi.admin.services.permission.actionProvider.getAll();
-  const contentTypesActions = allActions.filter(a => a.section === 'contentTypes');
-
-  const permissions = strapi.admin.services['content-type'].getPermissionsWithNestedFields(
-    contentTypesActions
-  );
-
-  const otherActions = allActions.filter(a => a.section !== 'contentTypes');
-  otherActions.forEach(action => {
-    if (action.subjects) {
-      const newPerms = action.subjects.map(subject =>
-        createPermission({ action: action.actionId, subject })
-      );
-      permissions.push(...newPerms);
-    } else {
-      permissions.push(createPermission({ action: action.actionId }));
-    }
-  });
-
-  await assign(superAdminRole.id, permissions);
 };
 
 module.exports = {
@@ -292,7 +192,6 @@ module.exports = {
   find,
   deleteByRolesIds,
   deleteByIds,
-  assign,
   sanitizePermission,
   findUserPermissions,
   actionProvider,
@@ -300,6 +199,5 @@ module.exports = {
   engine,
   conditionProvider,
   cleanPermissionInDatabase,
-  resetSuperAdminPermissions,
   ensureBoundPermissionsInDatabase,
 };
